create archived_blocks dir, update migrations
